% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_theory.R
\name{.circ_dist}
\alias{.circ_dist}
\title{Circular distance between hues, in degrees (0-180)}
\usage{
.circ_dist(a, b)
}
\description{
Circular distance between hues, in degrees (0-180)
}
\keyword{internal}
